+++ core/src/main/kotlin/no/elg/hex/screens/LevelSelectScreen.kt
@@ -93,11 +93,20 @@ class LevelSelectScreen : AbstractScreen(), ReloadableScreen {
 
       val preview = metadata.preview ?: continue
       batch.draw(preview, x, y, width, height)
-      if (Hex.debugStage || (Hex.args.debug && metadata.authorRoundsToBeat == Island.UNKNOWN_ROUNDS_TO_BEAT)) {
-        drawBox(x, y, width, height, Color.PURPLE)
-      }
-      if (Hex.args.debug && metadata.authorRoundsToBeat == Island.NEVER_BEATEN) {
-        drawBox(x, y, width, height, Color.RED)
+      if(Hex.debug) {
+        val color = if (metadata.authorRoundsToBeat == Island.UNKNOWN_ROUNDS_TO_BEAT) {
+          Color.PURPLE
+        }
+        else if (metadata.authorRoundsToBeat == Island.NEVER_BEATEN) {
+          Color.RED
+        }else if (metadata.authorRoundsToBeat == Int.MAX_VALUE) {
+          Color.GOLD
+        } else if (Hex.debugStage) {
+          NOT_SELECTED_COLOR
+        } else{
+          null
+        }
+        color?.let { drawBox(x, y, width, height, it) }
       }
     }
     batch.end()